refactor: split five files approaching the 500-line limit
Five engine/libpod files sat within ~12 lines of the 500-line cap. Each is split
along a cohesive seam by pure relocation, with unit tests moved alongside the
code they cover and no behavioural change:

- engine/volume_mounts: extract volume-spec string parsing into spec.rs
- engine/lifecycle: extract grace period, service filtering, and depends_on
  expansion into targets.rs
- libpod/client: extract urlencoded into encode.rs
- engine/container: extract name/path/link/config-hash resolution into resolve.rs
- engine/query: extract top, port, images, and attach_logs into inspect.rs

Closes #239
